Create fragment and transfer code from MainActivity
