+++ app/src/main/AndroidManifest.xml
@@ -11,10 +11,10 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.AppNotes">
         <activity
-            android:name=".ui.NoteActivity"
+            android:name=".ui.details.NoteActivity"
             android:exported="true" />
         <activity
-            android:name=".ui.MainActivity"
+            android:name=".ui.main.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
