+++ app/src/main/java/by/geekbrains/appnotes/ui/main/MainActivity.java
@@ -1,0 +1,17 @@
+package by.geekbrains.appnotes.ui.main;
+
+import android.os.Bundle;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import by.geekbrains.appnotes.R;
+
+public class MainActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+
+    }
+}
