+++ app/src/main/res/values/strings.xml
@@ -3,4 +3,6 @@
     <string name="hint_title_note_edit_text">Тема</string>
     <string name="hint_description_note_edit_text">Описание</string>
     <string name="text_save_note_button">Save note</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
